Suggested restructuring of flavor transformations * move transformation presets to __init__.py * move FlavorTransformation base class to base.py
